move predict_noise_variance_from_autos from utils to noise
